requirements.txt
#pandas==1.3
#numpy==1.25.2
faiss-cpu
pandas
numpy
gradio
huggingface_hub
fuzzywuzzy
python-Levenshtein